Pass dtype to gaussian_2d

+++ tests/_basis_util.py
@@ -39,8 +39,9 @@ class Steerable2DMixin:
         # Want sigma to be as large as possible without the Gaussian
         # spilling too much outside the central disk.
         sigma = self.L / 8
-        im1 = gaussian_2d(self.L, x0=x0, y0=y0, sigma_x=sigma, sigma_y=sigma)
-        im1 = im1.astype(self.dtype)
+        im1 = gaussian_2d(
+            self.L, x0=x0, y0=y0, sigma_x=sigma, sigma_y=sigma, dtype=self.dtype
+        )
 
         coef = self.basis.expand(im1)
         im2 = self.basis.evaluate(coef)
@@ -59,8 +60,7 @@ class Steerable2DMixin:
 
     def testIsotropic(self):
         sigma = self.L / 8
-        im = gaussian_2d(self.L, sigma_x=sigma, sigma_y=sigma)
-        im = im.astype(self.dtype)
+        im = gaussian_2d(self.L, sigma_x=sigma, sigma_y=sigma, dtype=self.dtype)
 
         coef = self.basis.expand(im)
 
@@ -80,8 +80,7 @@ class Steerable2DMixin:
         ell = 1
 
         sigma = self.L / 8
-        im = gaussian_2d(self.L, sigma_x=sigma, sigma_y=sigma)
-        im = im.astype(self.dtype)
+        im = gaussian_2d(self.L, sigma_x=sigma, sigma_y=sigma, dtype=self.dtype)
 
         g2d = grid_2d(self.L)
